Fix compilation errors by adding shared ModernUI classes and cleaning up duplicates
